Change core.styling into a package
Now parser, selector and all can be put in separate modules.
