+++ gaphor/core/styling/tests/test_selector.py
